refactor(mbx): split CLI tests by command
