+++ crates/mbx/src/cli/cargo_tests.rs
@@ -1,18 +1,6 @@
 use super::*;
 
 #[test]
-fn combined_budget_reserves_the_full_action_store_allowance() {
-    let retention = RetentionSettings {
-        target_max_bytes: Some(80),
-        target_max_age: None,
-        max_total_bytes: Some(100),
-    };
-
-    assert_eq!(target_budget(&retention, 70), Some(30));
-    assert_eq!(target_budget(&retention, 120), Some(0));
-}
-
-#[test]
 fn prefers_the_outermost_lockfile_as_the_workspace_root() {
     let directory = tempfile::tempdir().unwrap();
     let root = directory.path();
@@ -40,32 +28,6 @@ fn falls_back_to_a_manifest_without_a_lockfile() {
 fn falls_back_to_the_starting_directory() {
     let directory = tempfile::tempdir().unwrap();
     assert_eq!(workspace_root(directory.path()), directory.path());
-}
-
-#[test]
-/// A non-colocated Jujutsu checkout is a project boundary.
-fn discovers_a_jujutsu_project_root() {
-    let directory = tempfile::tempdir().unwrap();
-    let root = directory.path();
-    let nested = root.join("build").join("debug");
-    std::fs::create_dir_all(&nested).unwrap();
-    std::fs::create_dir(root.join(".jj")).unwrap();
-
-    assert_eq!(discover_project_root(&nested), root);
-}
-
-#[test]
-/// A nested Git checkout takes precedence over an enclosing Jujutsu checkout.
-fn discovers_a_nested_git_project_root() {
-    let directory = tempfile::tempdir().unwrap();
-    let outer = directory.path();
-    let inner = outer.join("vendor");
-    let nested = inner.join("src");
-    std::fs::create_dir_all(&nested).unwrap();
-    std::fs::create_dir(outer.join(".jj")).unwrap();
-    std::fs::create_dir(inner.join(".git")).unwrap();
-
-    assert_eq!(discover_project_root(&nested), inner);
 }
 
 #[test]
@@ -99,49 +61,6 @@ fn reads_the_roots_cargo_reports() {
             target_dir: PathBuf::from("/var/cache/shared-target"),
             target_dir_requested: false,
         }
-    );
-}
-
-#[test]
-fn cache_removal_uses_the_workspace_identity_reported_by_cargo() {
-    let directory = tempfile::tempdir().unwrap();
-    let requested = directory.path().join("workspace-alias");
-    std::fs::create_dir_all(&requested).unwrap();
-    std::fs::write(
-        requested.join("Cargo.toml"),
-        "[package]\nname = \"cache-remove-root\"\nversion = \"0.0.0\"\n",
-    )
-    .unwrap();
-    std::fs::create_dir(requested.join("src")).unwrap();
-    std::fs::write(requested.join("src/lib.rs"), "").unwrap();
-
-    let reported = cache_workspace_root(std::ffi::OsStr::new("cargo"), &requested);
-
-    let metadata = std::process::Command::new("cargo")
-        .args([
-            "metadata",
-            "--no-deps",
-            "--format-version",
-            "1",
-            "--manifest-path",
-        ])
-        .arg(requested.join("Cargo.toml"))
-        .output()
-        .unwrap();
-    assert!(metadata.status.success());
-    assert_eq!(
-        reported,
-        parse_cargo_roots(&metadata.stdout).unwrap().workspace_root
-    );
-}
-
-#[test]
-fn cache_removal_preserves_the_requested_spelling_when_cargo_cannot_resolve_it() {
-    let requested = PathBuf::from("/workspace/that/does/not/exist");
-
-    assert_eq!(
-        cache_workspace_root(std::ffi::OsStr::new("definitely-not-cargo"), &requested),
-        requested
     );
 }
 
@@ -424,243 +343,6 @@ fn a_command_line_config_include_may_set_the_target_dir() {
 }
 
 #[test]
-fn all_non_reserved_subcommands_are_forwarded_directly() {
-    for argv in [
-        ["mbx", "new", "--vcs", "none"],
-        ["mbx", "init", "--lib", "fixture"],
-        ["mbx", "command-added-later", "--future-flag", "value"],
-    ] {
-        let argv = argv.map(std::ffi::OsStr::new);
-        let cli = Cli::try_parse_from(&argv).unwrap();
-        let Commands::Cargo(arguments) = cli.command else {
-            panic!(
-                "{} should be treated as a cargo subcommand",
-                argv[1].to_string_lossy()
-            );
-        };
-        let expected = argv[1..]
-            .iter()
-            .map(|argument| argument.to_string_lossy().into_owned())
-            .collect::<Vec<_>>();
-        assert_eq!(arguments, expected);
-    }
-}
-
-#[test]
-fn a_toolchain_in_front_reaches_cargo_as_it_was_typed() {
-    let argv = ["mbx", "+1.91", "check", "--locked"].map(std::ffi::OsStr::new);
-
-    let cli = Cli::try_parse_from(&argv).unwrap();
-
-    assert_eq!(cli.toolchain.as_deref(), Some("1.91"));
-    let Commands::Cargo(arguments) = cli.command else {
-        panic!("check should still be a cargo subcommand");
-    };
-    // Read by mbx, and then handed back: cargo is the rustup shim, and the
-    // command line is how it is told which toolchain to run.
-    assert_eq!(arguments, ["check", "--locked"]);
-    assert_eq!(
-        with_toolchain(Some("1.91"), arguments),
-        ["+1.91", "check", "--locked"]
-    );
-}
-
-#[test]
-fn a_command_that_names_no_toolchain_is_forwarded_untouched() {
-    let argv = ["mbx", "check", "--locked"].map(std::ffi::OsStr::new);
-
-    let cli = Cli::try_parse_from(&argv).unwrap();
-
-    assert_eq!(cli.toolchain, None);
-    let Commands::Cargo(arguments) = cli.command else {
-        panic!("check should still be a cargo subcommand");
-    };
-    assert_eq!(with_toolchain(None, arguments), ["check", "--locked"]);
-}
-
-#[test]
-fn a_plus_word_inside_a_cargo_command_belongs_to_cargo() {
-    // The sigil classifies the word in front of the command. Past that point the
-    // line is the cargo command's own, and a `+` in it means whatever cargo says
-    // it means — including one held by a flag, or handed to the built program.
-    for argv in [
-        ["mbx", "test", "+weird", "--lib"],
-        ["mbx", "build", "--features", "+simd"],
-        ["mbx", "run", "--", "+5"],
-    ] {
-        let owned = argv.map(std::ffi::OsStr::new);
-        let cli = Cli::try_parse_from(&owned).unwrap();
-        assert_eq!(cli.toolchain, None, "{argv:?} names no toolchain to mbx");
-        let Commands::Cargo(arguments) = cli.command else {
-            panic!("{argv:?} should be a cargo subcommand");
-        };
-        assert_eq!(arguments, argv[1..]);
-    }
-}
-
-#[test]
-fn a_toolchain_selects_one_for_every_command_that_reaches_a_compiler() {
-    let explain = ["mbx", "+nightly", "explain", "clippy"].map(std::ffi::OsStr::new);
-
-    for argv in [
-        vec!["mbx", "+nightly", "build"],
-        vec!["mbx", "+nightly", "explain", "clippy"],
-        vec!["mbx", "+nightly", "doctor"],
-    ] {
-        let owned = argv.iter().map(std::ffi::OsStr::new).collect::<Vec<_>>();
-        let cli = Cli::try_parse_from(&owned).unwrap();
-        assert_eq!(cli.toolchain.as_deref(), Some("nightly"), "{argv:?}");
-        assert_eq!(compiles_nothing(&cli.command), None, "{argv:?}");
-    }
-
-    let cli = Cli::try_parse_from(&explain).unwrap();
-    let Commands::Explain(args) = cli.command else {
-        panic!("explain should be reserved by mbx");
-    };
-    assert_eq!(
-        with_toolchain(Some("nightly"), args.arguments()),
-        ["+nightly", "clippy"]
-    );
-}
-
-#[test]
-fn the_words_before_the_subcommand_are_mbx_s_with_a_toolchain_or_without() {
-    // mbx reads what comes before the subcommand, so `mbx -q build` has always
-    // been an unknown flag rather than a quiet cargo build. A toolchain used to
-    // switch that off by accident — `+stable` matched no command mbx knew, so
-    // the rest of the line went to cargo unexamined — and now it does not.
-    // Cargo's own globals still work where cargo takes them after the
-    // subcommand, as `mbx build -q` does.
-    for argv in [
-        vec!["mbx", "-q", "build"],
-        vec!["mbx", "+stable", "-q", "build"],
-        vec!["mbx", "--offline", "build"],
-        vec!["mbx", "+stable", "--offline", "build"],
-    ] {
-        let owned = argv.iter().map(std::ffi::OsStr::new).collect::<Vec<_>>();
-        assert!(
-            Cli::try_parse_from(&owned).is_err(),
-            "{argv:?} should be read by mbx, not handed over whole"
-        );
-    }
-
-    let owned = ["mbx", "build", "-q"].map(std::ffi::OsStr::new);
-    let cli = Cli::try_parse_from(&owned).unwrap();
-    let Commands::Cargo(arguments) = cli.command else {
-        panic!("build should be a cargo subcommand");
-    };
-    assert_eq!(arguments, ["build", "-q"]);
-}
-
-#[test]
-fn a_toolchain_is_refused_where_no_compiler_would_see_it() {
-    for (argv, command) in [
-        (vec!["mbx", "+1.91", "gc"], "gc"),
-        (vec!["mbx", "+1.91", "cache", "dir"], "cache"),
-        (vec!["mbx", "+1.91", "tui"], "tui"),
-        (vec!["mbx", "+1.91", "exec", "make"], "exec"),
-        (vec!["mbx", "+1.91", "setup"], "setup"),
-    ] {
-        let owned = argv.iter().map(std::ffi::OsStr::new).collect::<Vec<_>>();
-        let cli = Cli::try_parse_from(&owned).unwrap();
-        // Refused rather than ignored: naming a toolchain for a command that
-        // compiles nothing is a misunderstanding worth reporting.
-        assert_eq!(compiles_nothing(&cli.command), Some(command), "{argv:?}");
-    }
-}
-
-#[test]
-fn mbx_commands_still_take_precedence() {
-    let argv = ["mbx", "gc", "--max-size", "1GiB"].map(std::ffi::OsStr::new);
-    let cli = Cli::try_parse_from(&argv).unwrap();
-    assert!(matches!(cli.command, Commands::Gc(_)));
-}
-
-#[test]
-fn explain_forwards_cargo_flags_and_the_rustc_separator() {
-    let argv = [
-        "mbx",
-        "explain",
-        "clippy",
-        "--workspace",
-        "--",
-        "-D",
-        "warnings",
-    ]
-    .map(std::ffi::OsStr::new);
-    let cli = Cli::try_parse_from(&argv).unwrap();
-    let Commands::Explain(arguments) = cli.command else {
-        panic!("explain should be reserved by mbx");
-    };
-    assert_eq!(
-        arguments.arguments(),
-        ["clippy", "--workspace", "--", "-D", "warnings"]
-    );
-}
-
-#[test]
-fn prefetch_preserves_cargo_flags_and_the_rustc_separator() {
-    let argv = [
-        "mbx",
-        "prefetch",
-        "test",
-        "--workspace",
-        "--",
-        "--nocapture",
-    ]
-    .map(std::ffi::OsStr::new);
-
-    let cli = Cli::try_parse_from(&argv).unwrap();
-
-    let Commands::Prefetch(args) = cli.command else {
-        panic!("prefetch should remain an mbx command");
-    };
-    assert_eq!(args.cargo_args, ["test", "--workspace", "--nocapture"]);
-    let original = argv
-        .iter()
-        .map(|argument| argument.to_os_string())
-        .collect::<Vec<_>>();
-    assert_eq!(
-        original_prefetch_arguments(&original).unwrap(),
-        ["test", "--workspace", "--", "--nocapture"]
-    );
-}
-
-#[test]
-fn prefetch_accepts_a_read_capable_remote() {
-    let directory = tempfile::tempdir().unwrap();
-    let mut config = managed_target_config(directory.path());
-    config.remote.url = Some("https://cache.example.test".into());
-    config.remote.mode = mbx_cache_core::RemoteCacheMode::ReadOnly;
-
-    assert!(validate_prefetch_config(&config).is_ok());
-}
-
-#[test]
-fn cli_exposes_its_usage_spec() {
-    let spec = Cli::to_kdl();
-    assert!(spec.contains("external_subcommand #true"));
-    // Still declared, so it keeps working and stays out of help and
-    // completions rather than being removed from under anyone using it.
-    assert!(spec.contains("cmd setup"));
-    assert!(
-        spec.contains("hide=#true"),
-        "setup should be hidden from help: {spec}"
-    );
-    assert!(
-        spec.contains("sigil=+"),
-        "the toolchain argument should be classified by its sigil: {spec}"
-    );
-    assert!(spec.contains("cmd explain"));
-    assert!(spec.contains("cmd doctor"));
-    assert!(spec.contains("cmd gc"));
-    assert!(spec.contains("cmd cache"));
-    assert!(spec.contains("config {"));
-    assert!(spec.contains(r#"prop "gc.max_size""#));
-    assert!(spec.contains(r#"env "MBX_GC_MAX_SIZE""#));
-}
-
-#[test]
 fn cargo_help_does_not_trigger_target_migration() {
     assert!(cargo_help_requested(&["build".into(), "--help".into()]));
     assert!(cargo_help_requested(&["help".into(), "build".into()]));
@@ -760,7 +442,7 @@ fn reasons_read_as_prose() {
     );
 }
 
-fn managed_target_config(root: &Path) -> Config {
+pub(super) fn managed_target_config(root: &Path) -> Config {
     Config {
         cache_dir: root.join("cache"),
         stats_report: None,
@@ -818,159 +500,4 @@ fn declining_the_target_prompt_preserves_outputs() {
 
     assert!(!accepted);
     assert!(target_dir.join("artifact").is_file());
-}
-
-#[test]
-fn setup_preserves_cargo_configuration_and_installs_the_wrapper() {
-    let directory = tempfile::tempdir().unwrap();
-    let executable = directory
-        .path()
-        .join(if cfg!(windows) { "mbx.exe" } else { "mbx" });
-    std::fs::write(&executable, b"mbx binary").unwrap();
-    let install = directory.path().join("data/bin");
-    let config = directory.path().join("cargo/config.toml");
-    std::fs::create_dir_all(config.parent().unwrap()).unwrap();
-    std::fs::write(&config, "# keep me\n[net]\noffline = true\n").unwrap();
-
-    setup_at(&executable, &install, &config).unwrap();
-
-    let written = std::fs::read_to_string(&config).unwrap();
-    assert!(written.contains("# keep me"));
-    let document = written.parse::<toml_edit::DocumentMut>().unwrap();
-    assert_eq!(document["net"]["offline"].as_bool(), Some(true));
-    let wrapper = document["build"]["rustc-wrapper"].as_str().unwrap();
-    assert!(Path::new(wrapper).is_file());
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::PermissionsExt as _;
-
-        assert_ne!(
-            std::fs::metadata(wrapper).unwrap().permissions().mode() & 0o100,
-            0
-        );
-    }
-}
-
-#[test]
-fn setup_never_displaces_an_existing_wrapper() {
-    let directory = tempfile::tempdir().unwrap();
-    let executable = directory.path().join("mbx");
-    std::fs::write(&executable, b"mbx binary").unwrap();
-    let install = directory.path().join("data/bin");
-    let config = directory.path().join("cargo/config.toml");
-    std::fs::create_dir_all(config.parent().unwrap()).unwrap();
-    let original = "[build]\nrustc-wrapper = \"sccache\"\n";
-    std::fs::write(&config, original).unwrap();
-
-    setup_at(&executable, &install, &config).unwrap();
-
-    assert_eq!(std::fs::read_to_string(config).unwrap(), original);
-    assert!(!install.exists());
-}
-
-#[test]
-fn setup_flags_are_mutually_exclusive() {
-    let args = SetupArgs {
-        status: true,
-        update: true,
-        uninstall: false,
-    };
-    assert!(args.action().is_err());
-    assert_eq!(
-        SetupArgs {
-            status: false,
-            update: false,
-            uninstall: true,
-        }
-        .action()
-        .unwrap(),
-        SetupAction::Uninstall
-    );
-}
-
-#[test]
-fn setup_status_detects_and_update_refreshes_a_stale_wrapper() {
-    let directory = tempfile::tempdir().unwrap();
-    let executable = directory.path().join("mbx");
-    std::fs::write(&executable, b"first mbx binary").unwrap();
-    let install = directory.path().join("data/bin");
-    let config = directory.path().join("cargo/config.toml");
-
-    setup_at_action(&executable, &install, &config, SetupAction::Install).unwrap();
-    assert_eq!(
-        setup_at_action(&executable, &install, &config, SetupAction::Status).unwrap(),
-        ExitCode::SUCCESS
-    );
-
-    std::fs::remove_file(&executable).unwrap();
-    std::fs::write(&executable, b"new mbx binary with another size").unwrap();
-    assert_eq!(
-        setup_at_action(&executable, &install, &config, SetupAction::Status).unwrap(),
-        ExitCode::FAILURE
-    );
-    setup_at_action(&executable, &install, &config, SetupAction::Update).unwrap();
-    let shim = install.join(if cfg!(windows) {
-        "mbx-rustc.exe"
-    } else {
-        "mbx-rustc"
-    });
-    assert!(same_file_contents(&executable, &shim).unwrap());
-}
-
-#[test]
-fn setup_uninstall_removes_only_mbx_configuration() {
-    let directory = tempfile::tempdir().unwrap();
-    let executable = directory.path().join("mbx");
-    std::fs::write(&executable, b"mbx binary").unwrap();
-    let install = directory.path().join("data/bin");
-    let config = directory.path().join("cargo/config.toml");
-    std::fs::create_dir_all(config.parent().unwrap()).unwrap();
-    std::fs::write(&config, "# keep me\n[net]\noffline = true\n").unwrap();
-
-    setup_at_action(&executable, &install, &config, SetupAction::Install).unwrap();
-    let shim = install.join(if cfg!(windows) {
-        "mbx-rustc.exe"
-    } else {
-        "mbx-rustc"
-    });
-    assert!(shim.is_file());
-    setup_at_action(&executable, &install, &config, SetupAction::Uninstall).unwrap();
-
-    let contents = std::fs::read_to_string(&config).unwrap();
-    assert!(contents.contains("# keep me"));
-    assert!(contents.contains("offline = true"));
-    assert!(!contents.contains("rustc-wrapper"));
-    assert!(!shim.exists());
-    setup_at_action(&executable, &install, &config, SetupAction::Uninstall).unwrap();
-}
-
-#[test]
-fn setup_update_requires_an_existing_installation() {
-    let directory = tempfile::tempdir().unwrap();
-    let executable = directory.path().join("mbx");
-    std::fs::write(&executable, b"mbx binary").unwrap();
-    let error = setup_at_action(
-        &executable,
-        &directory.path().join("data/bin"),
-        &directory.path().join("cargo/config.toml"),
-        SetupAction::Update,
-    )
-    .unwrap_err();
-    assert!(error.to_string().contains("not installed"));
-
-    let config = directory.path().join("cargo/config.toml");
-    std::fs::create_dir_all(config.parent().unwrap()).unwrap();
-    std::fs::write(&config, "[build]\nrustc-wrapper = \"sccache\"\n").unwrap();
-    let error = setup_at_action(
-        &executable,
-        &directory.path().join("data/bin"),
-        &config,
-        SetupAction::Update,
-    )
-    .unwrap_err();
-    assert!(
-        error
-            .to_string()
-            .contains("refusing to replace another wrapper")
-    );
 }
